+++ projects/miopen/src/include/miopen/conv/invokers/impl_gemm_dynamic.hpp
@@ -49,21 +49,21 @@ ComputeDynamicIGemmForwardKernelArgs<int>(const ProblemDescription& problem, con
 {
     std::vector<OpKernelArg> opArgs;
     // clang-format off
-    int hi          = problem.GetInHeight();
-    int wi          = problem.GetInWidth();
-    int n           = problem.GetInBatchSize();
-    int k           = problem.GetOutChannels();
-    int c           = problem.GetInChannels();
-    int ho          = problem.GetOutHeight();
-    int wo          = problem.GetOutWidth();
+    int hi          = int(problem.GetInHeight());
+    int wi          = int(problem.GetInWidth());
+    int n           = int(problem.GetInBatchSize());
+    int k           = int(problem.GetOutChannels());
+    int c           = int(problem.GetInChannels());
+    int ho          = int(problem.GetOutHeight());
+    int wo          = int(problem.GetOutWidth());
     int stride_h    = problem.GetKernelStrideH();
     int stride_w    = problem.GetKernelStrideW();
     int dilation_h  = problem.GetDilationH();
     int dilation_w  = problem.GetDilationW();
     int pad_h       = problem.GetPadH();
     int pad_w       = problem.GetPadW();
-    int y           = problem.GetWeightsHeight();
-    int x           = problem.GetWeightsWidth();
+    int y           = int(problem.GetWeightsHeight());
+    int x           = int(problem.GetWeightsWidth());
     int pack0       = cfg;
     // clang-format on
 
@@ -97,21 +97,21 @@ ComputeDynamicIGemmForwardKernelArgs<solver::TunableImplicitGemmGTCDynamic_t>(
 {
     std::vector<OpKernelArg> opArgs;
     // clang-format off
-    int hi          = problem.GetInHeight();
-    int wi          = problem.GetInWidth();
-    int n           = problem.GetInBatchSize();
-    int k           = problem.GetOutChannels();
-    int c           = problem.GetInChannels();
-    int ho          = problem.GetOutHeight();
-    int wo          = problem.GetOutWidth();
+    int hi          = int(problem.GetInHeight());
+    int wi          = int(problem.GetInWidth());
+    int n           = int(problem.GetInBatchSize());
+    int k           = int(problem.GetOutChannels());
+    int c           = int(problem.GetInChannels());
+    int ho          = int(problem.GetOutHeight());
+    int wo          = int(problem.GetOutWidth());
     int stride_h    = problem.GetKernelStrideH();
     int stride_w    = problem.GetKernelStrideW();
     int dilation_h  = problem.GetDilationH();
     int dilation_w  = problem.GetDilationW();
     int pad_h       = problem.GetPadH();
     int pad_w       = problem.GetPadW();
-    int y           = problem.GetWeightsHeight();
-    int x           = problem.GetWeightsWidth();
+    int y           = int(problem.GetWeightsHeight());
+    int x           = int(problem.GetWeightsWidth());
     int group       = problem.GetGroupCount();
     int pack0       = 0;
     // clang-format on
@@ -124,19 +124,19 @@ ComputeDynamicIGemmForwardKernelArgs<solver::TunableImplicitGemmGTCDynamic_t>(
         nxe == 0 ? (ho * wo) : ((ho * wo + nxb - 1) / nxb) * nxb; // pad to nxb modulo when nxe != 0
 
     // init magic division parameters
-    uint32_t nb_n0          = cfg.tensor_b_cluster_lengths[2] * cfg.tensor_b_thread_lengths[2];
-    uint32_t nb_n1b         = cfg.tensor_b_cluster_lengths[3] * cfg.tensor_b_thread_lengths[3];
-    uint32_t unmerge_sub_n  = cfg.gemm_n_per_block / nxb;
+    uint32_t nb_n0  = uint32_t(cfg.tensor_b_cluster_lengths[2] * cfg.tensor_b_thread_lengths[2]);
+    uint32_t nb_n1b = uint32_t(cfg.tensor_b_cluster_lengths[3] * cfg.tensor_b_thread_lengths[3]);
+    uint32_t unmerge_sub_n  = uint32_t(cfg.gemm_n_per_block / nxb);
     uint32_t unmerge_sub_n1 = unmerge_sub_n / nb_n0;
 
-    magic_div_u32_t mdiv_0 = magic_div_u32_gen(gemm_m / cfg.gemm_m_per_block);
-    magic_div_u32_t mdiv_1 = magic_div_u32_gen(b * unmerge_sub_n1 / nb_n1b);
-    magic_div_u32_t mdiv_2 = magic_div_u32_gen(y * x);
-    magic_div_u32_t mdiv_3 = magic_div_u32_gen(x);
-    magic_div_u32_t mdiv_4 = magic_div_u32_gen(b);
-    magic_div_u32_t mdiv_5 = magic_div_u32_gen(wo);
-    magic_div_u32_t mdiv_6 =
-        magic_div_u32_gen((n * b * (gemm_m)) / (cfg.gemm_m_per_block * cfg.gemm_n_per_block));
+    magic_div_u32_t mdiv_0 = magic_div_u32_gen(uint32_t(gemm_m / cfg.gemm_m_per_block));
+    magic_div_u32_t mdiv_1 = magic_div_u32_gen(uint32_t(b) * unmerge_sub_n1 / nb_n1b);
+    magic_div_u32_t mdiv_2 = magic_div_u32_gen(uint32_t(y * x));
+    magic_div_u32_t mdiv_3 = magic_div_u32_gen(uint32_t(x));
+    magic_div_u32_t mdiv_4 = magic_div_u32_gen(uint32_t(b));
+    magic_div_u32_t mdiv_5 = magic_div_u32_gen(uint32_t(wo));
+    magic_div_u32_t mdiv_6 = magic_div_u32_gen(
+        uint32_t((n * b * gemm_m) / (cfg.gemm_m_per_block * cfg.gemm_n_per_block)));
 
     uint32_t magic_0 = mdiv_0.magic;
     uint32_t magic_1 = mdiv_1.magic;
